1. 修改首页多个页面UI 2. 修复搜索失败的重试错误

+++ app/src/main/java/com/oplay/giftcool/model/data/resp/IndexGiftNew.java
@@ -83,7 +83,7 @@ public class IndexGiftNew implements Serializable {
     public int giftType;
     // 原价
     @SerializedName("original_price")
-    public int originPrice;
+    public float originPrice;
     // 剩余开抢时间，单位:s
     @SerializedName("remain_time")
     public int remainStartTime;
